examples/TreeView: 将计算机视图相关的内容封装在ComputerView类中。
