+++ examples/TreeView/MainForm.cpp
@@ -1,5 +1,6 @@
 #include "MainForm.h"
 #include "MainThread.h"
+#include "ComputerView.h"
 
 MainForm::MainForm():
     m_fileList(this),
@@ -11,17 +12,12 @@ MainForm::MainForm():
     m_pBtnBack(nullptr),
     m_bCanAddBackForward(true),
     m_pTreeNode(nullptr),
-    m_pTabBox(nullptr),
-    m_pComputerListCtrl(nullptr),
-    m_nRemoveIconCallbackId(0)
+    m_pTabBox(nullptr)
 {
 }
 
 MainForm::~MainForm()
 {
-    ui::GlobalManager::Instance().Icon().DetachRemoveIconEvent(m_nRemoveIconCallbackId);
-    ClearDiskInfoList(m_diskInfoList);
-    m_diskInfoList.clear();
 }
 
 DString MainForm::GetSkinFolder()
@@ -36,9 +32,6 @@ DString MainForm::GetSkinFile()
 
 void MainForm::OnInitWindow()
 {
-    //挂载图标删除事件
-    m_nRemoveIconCallbackId = ui::GlobalManager::Instance().Icon().AttachRemoveIconEvent(ui::UiBind(&MainForm::OnRemoveIcon, this, std::placeholders::_1));
-
     m_pTree = dynamic_cast<ui::DirectoryTree*>(FindControl(_T("tree")));
     ASSERT(m_pTree != nullptr);
     if (m_pTree == nullptr) {
@@ -49,17 +42,8 @@ void MainForm::OnInitWindow()
         m_pAddressBar->AttachReturn(UiBind(&MainForm::OnAddressBarReturn, this, std::placeholders::_1));
     }
     m_pTabBox = dynamic_cast<ui::TabBox*>(FindControl(_T("main_view_tab_box")));
-    m_pComputerListCtrl = dynamic_cast<ui::ListCtrl*>(FindControl(_T("computer_view")));
-    InitializeComputerViewHeader();
-    if (m_pComputerListCtrl != nullptr) {
-        //关联图片列表
-        ui::ImageListPtr pImageList = std::make_shared<ui::ImageList>();
-        pImageList->SetImageSize(ui::UiSize(16, 16), Dpi(), true);
-        m_pComputerListCtrl->SetImageList(ui::ListCtrlType::Report, pImageList);//该指针的资源生命周期由ListCtrl内部管理
-        
-        //挂载列表项鼠标双击事件
-        m_pComputerListCtrl->AttachDoubleClick(UiBind(&MainForm::OnComuterViewDoubleClick, this, std::placeholders::_1));
-    }
+    ui::ListCtrl* pComputerListCtrl = dynamic_cast<ui::ListCtrl*>(FindControl(_T("computer_view")));
+    m_pComputerView = std::make_unique<ComputerView>(this, pComputerListCtrl);
     m_pListBox = dynamic_cast<ui::VirtualListBox*>(FindControl(_T("file_view")));
     ASSERT(m_pListBox != nullptr);
     if (m_pListBox != nullptr) {
@@ -215,22 +199,26 @@ void MainForm::OnShowMyComputerContents(ui::TreeNode* pTreeNode,
                                         const std::vector<ui::DirectoryTree::DiskInfo>& diskInfoList)
 {
     ui::GlobalManager::Instance().AssertUIThread();
+    if ((pTreeNode == nullptr) || (m_pTree == nullptr) || !m_pTree->IsValidTreeNode(pTreeNode)) {
+        if (m_pComputerView != nullptr) {
+            m_pComputerView->ClearDiskInfoList(diskInfoList);
+        }
+        return;
+    }
     if (m_pTabBox != nullptr) {
         m_pTabBox->SelectItem((size_t)FormViewType::kComputerView);
-    }
-    if ((pTreeNode == nullptr) || (m_pTree == nullptr)) {
-        ClearDiskInfoList(diskInfoList);
-        return;
-    }
-    if (!m_pTree->IsValidTreeNode(pTreeNode)) {
-        ClearDiskInfoList(diskInfoList);
-        return;
     }
     if (m_pAddressBar != nullptr) {
         m_pAddressBar->SetText(_T(""));
     }
-    SetShowTreeNode(pTreeNode);
-    FillMyComputerContents(diskInfoList);
+
+    //显示计算机视图的内容
+    if (m_pComputerView != nullptr) {
+        m_pComputerView->ShowMyComputerContents(diskInfoList);
+    }   
+
+    //保存显示的树节点
+    SetShowTreeNode(pTreeNode);   
 
     //更新界面状态
     UpdateCommandUI();
@@ -407,397 +395,5 @@ void MainForm::UpdateCommandUI()
     }
     if (m_pBtnForward != nullptr) {
         m_pBtnForward->SetEnabled(bEnableForward);
-    }
-}
-
-void MainForm::InitializeComputerViewHeader()
-{
-    if (m_pComputerListCtrl == nullptr) {
-        return;
-    }
-    ui::ListCtrlHeaderItem* pHeaderItem = nullptr;
-#ifdef DUILIB_BUILD_FOR_WIN
-    ui::ListCtrlColumn columnInfo;
-    columnInfo.text = _T("名称");
-    columnInfo.nColumnWidth = 200;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kName] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("磁盘类型");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kType] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("分区类型");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kPartitionType] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("总大小");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kTotalSpace] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("可用空间");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kFreeSpace] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("已用");
-    columnInfo.nColumnWidth = 80;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kUsedPercent] = pHeaderItem->GetColumnId();
-#else
-    ui::ListCtrlColumn columnInfo;
-    columnInfo.text = _T("文件系统");
-    columnInfo.nColumnWidth = 200;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kName] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("设备类型");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kType] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("分区类型");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kPartitionType] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("总大小");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kTotalSpace] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("可用空间");
-    columnInfo.nColumnWidth = 120;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kFreeSpace] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("已用");
-    columnInfo.nColumnWidth = 80;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kUsedPercent] = pHeaderItem->GetColumnId();
-
-    columnInfo.text = _T("挂载点");
-    columnInfo.nColumnWidth = 200;
-    pHeaderItem = m_pComputerListCtrl->InsertColumn(-1, columnInfo);
-    ASSERT(pHeaderItem != nullptr);
-    m_columnIdMap[ComputerViewColumn::kMountOn] = pHeaderItem->GetColumnId();
-#endif
-}
-
-void MainForm::FillMyComputerContents(const std::vector<ui::DirectoryTree::DiskInfo>& diskInfoList)
-{
-    ui::GlobalManager::Instance().AssertUIThread();
-    ClearDiskInfoList(m_diskInfoList);
-    m_diskInfoList = diskInfoList;
-    if ((m_pComputerListCtrl == nullptr)){
-        return;
-    }
-    m_pComputerListCtrl->DeleteAllDataItems();
-    ui::ImageListPtr pImageList;
-    if (m_pComputerListCtrl != nullptr) {
-        pImageList = m_pComputerListCtrl->GetImageList(ui::ListCtrlType::Report);
-    }
-    if (pImageList != nullptr) {
-        pImageList->Clear();
-    }
-
-    ui::ListCtrlSubItemData itemData;
-    size_t nItemIndex = 0;
-    size_t nColumnId = 0;
-#ifdef DUILIB_BUILD_FOR_WIN
-    for (size_t nIndex = 0; nIndex < diskInfoList.size(); ++nIndex) {
-        const ui::DirectoryTree::DiskInfo& diskInfo = diskInfoList[nIndex];   
-        nItemIndex = m_pComputerListCtrl->AddDataItem(itemData);
-        if (ui::Box::IsValidItemIndex(nItemIndex)) {
-            //记录关联关系
-            m_pComputerListCtrl->SetDataItemUserData(nItemIndex, nIndex);
-            //设置图标
-            if (pImageList != nullptr) {
-                DString iconString = ui::GlobalManager::Instance().Icon().GetIconString(diskInfo.m_nIconID);
-                if (!iconString.empty()) {
-                    int32_t nImageId = pImageList->AddImageString(iconString, Dpi());
-                    m_iconToImageMap[diskInfo.m_nIconID] = nImageId;
-                    m_pComputerListCtrl->SetDataItemImageId(nItemIndex, nImageId);
-                }
-            }
-
-            ui::ListCtrlSubItemData subItemData;
-            subItemData.nTextFormat = ui::DrawStringFormat::TEXT_LEFT | ui::DrawStringFormat::TEXT_VCENTER;
-
-            subItemData.text = diskInfo.m_displayName;
-            nColumnId = GetColumnId(ComputerViewColumn::kName);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //名称
-            //设置"名称"列的排序方式
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, nIndex);
-            m_pComputerListCtrl->SetSubItemUserDataSById(nItemIndex, nColumnId, diskInfo.m_filePath.ToString());
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.nTextFormat = ui::DrawStringFormat::TEXT_CENTER | ui::DrawStringFormat::TEXT_VCENTER;
-
-            subItemData.text = diskInfo.m_volumeType;
-            nColumnId = GetColumnId(ComputerViewColumn::kType);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //磁盘类型
-
-            subItemData.text = diskInfo.m_fileSystem;
-            nColumnId = GetColumnId(ComputerViewColumn::kPartitionType);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //分区类型
-
-            subItemData.text = FormatDiskSpace(diskInfo.m_totalBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kTotalSpace);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //总大小
-
-            //设置"总大小"列的排序方式
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, diskInfo.m_totalBytes);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.text = FormatDiskSpace(diskInfo.m_freeBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kFreeSpace);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //可用空间
-
-            //设置"可用空间"列的排序方式
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, diskInfo.m_freeBytes);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.text = FormatUsedPercent(diskInfo.m_totalBytes, diskInfo.m_freeBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kUsedPercent);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //已用百分比
-
-            //设置"已用百分比"列的排序方式
-            uint64_t nUsedPercent = 100 * 1000;
-            if (diskInfo.m_totalBytes != 0) {
-                nUsedPercent = (diskInfo.m_totalBytes - diskInfo.m_freeBytes) * 1000 / diskInfo.m_totalBytes;
-            }
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, (size_t)nUsedPercent);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-        }
-    }
-#else
-    for (size_t nIndex = 0; nIndex < diskInfoList.size(); ++nIndex) {
-        const ui::DirectoryTree::DiskInfo& diskInfo = diskInfoList[nIndex];        
-        nItemIndex = m_pComputerListCtrl->AddDataItem(itemData);
-        if (ui::Box::IsValidItemIndex(nItemIndex)) {
-            //记录关联关系
-            m_pComputerListCtrl->SetDataItemUserData(nItemIndex, nIndex);
-
-            //设置图标
-            if (pImageList != nullptr) {
-                DString iconString = ui::GlobalManager::Instance().Icon().GetIconString(diskInfo.m_nIconID);
-                if (!iconString.empty()) {
-                    int32_t nImageId = pImageList->AddImageString(iconString, Dpi());
-                    m_iconToImageMap[diskInfo.m_nIconID] = nImageId;
-                    m_pComputerListCtrl->SetDataItemImageId(nItemIndex, nImageId);
-                }
-            }
-
-            ui::ListCtrlSubItemData subItemData;
-            subItemData.nTextFormat = ui::DrawStringFormat::TEXT_LEFT | ui::DrawStringFormat::TEXT_VCENTER;
-            subItemData.text = diskInfo.m_displayName;
-            nColumnId = GetColumnId(ComputerViewColumn::kName);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //文件系统
-
-            subItemData.nTextFormat = ui::DrawStringFormat::TEXT_CENTER | ui::DrawStringFormat::TEXT_VCENTER;
-
-            subItemData.text = GetDeviceTypeString(diskInfo.m_deviceType);
-            nColumnId = GetColumnId(ComputerViewColumn::kType);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //设备类型
-
-            subItemData.text = diskInfo.m_fileSystem;
-            nColumnId = GetColumnId(ComputerViewColumn::kPartitionType);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //分区类型
-
-            subItemData.text = FormatDiskSpace(diskInfo.m_totalBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kTotalSpace);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //总大小
-
-            //设置"总大小"列的排序方式
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, diskInfo.m_totalBytes);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.text = FormatDiskSpace(diskInfo.m_freeBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kFreeSpace);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //可用空间
-
-            //设置"可用空间"列的排序方式
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, diskInfo.m_freeBytes);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.text = FormatUsedPercent(diskInfo.m_totalBytes, diskInfo.m_freeBytes);
-            nColumnId = GetColumnId(ComputerViewColumn::kUsedPercent);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //已用百分比
-
-            //设置"已用百分比"列的排序方式
-            uint64_t nUsedPercent = 100 * 1000;
-            if (diskInfo.m_totalBytes != 0) {
-                nUsedPercent = (diskInfo.m_totalBytes - diskInfo.m_freeBytes) * 1000 / diskInfo.m_totalBytes;
-            }
-            m_pComputerListCtrl->SetSubItemUserDataNById(nItemIndex, nColumnId, (size_t)nUsedPercent);
-            m_pComputerListCtrl->SetColumnSortFlagById(nColumnId, ui::ListCtrlSubItemSortFlag::kSortByUserDataN);
-
-            subItemData.text = diskInfo.m_mountOn;
-            subItemData.nTextFormat = ui::DrawStringFormat::TEXT_LEFT | ui::DrawStringFormat::TEXT_VCENTER;
-            nColumnId = GetColumnId(ComputerViewColumn::kMountOn);
-            m_pComputerListCtrl->SetSubItemDataById(nItemIndex, nColumnId, subItemData); //挂载点
-        }
-    }
-#endif
-}
-
-#ifndef DUILIB_BUILD_FOR_WIN
-DString MainForm::GetDeviceTypeString(ui::DirectoryTree::DeviceType deviceType) const
-{
-    DString deviceTypeString = _T("未知");
-    switch (deviceType) {
-    case ui::DirectoryTree::DeviceType::HDD:
-        deviceTypeString = _T("机械硬盘");
-        break;
-    case ui::DirectoryTree::DeviceType::SSD:
-        deviceTypeString = _T("SATA固态硬盘");
-        break;
-    case ui::DirectoryTree::DeviceType::NVME:
-        deviceTypeString = _T("NVMe固态硬盘");
-        break;
-    case ui::DirectoryTree::DeviceType::USB:
-        deviceTypeString = _T("USB存储");
-        break;
-    case ui::DirectoryTree::DeviceType::SD_CARD:
-        deviceTypeString = _T("SD卡");
-        break;
-    case ui::DirectoryTree::DeviceType::CDROM:
-        deviceTypeString = _T("CD/DVD");
-        break;
-    case ui::DirectoryTree::DeviceType::LOOP:
-        deviceTypeString = _T("LOOP虚拟存储");
-        break;
-    case ui::DirectoryTree::DeviceType::VIRT_DISK:
-        deviceTypeString = _T("虚拟存储");
-        break;
-    case ui::DirectoryTree::DeviceType::RAMDISK:
-        deviceTypeString = _T("内存盘");
-        break;
-    case ui::DirectoryTree::DeviceType::NFS:
-        deviceTypeString = _T("NFS");
-        break;
-    case ui::DirectoryTree::DeviceType::SHARE:
-        deviceTypeString = _T("共享文件夹");
-        break;
-    default:
-        break;
-    }
-    return deviceTypeString;
-}
-#endif
-
-DString MainForm::FormatDiskSpace(uint64_t nSpace) const
-{
-    DString value;
-    if (nSpace > 1 * 1024 * 1024 * 1024) {
-        //GB
-        double total_gb = static_cast<double>(nSpace) / (1024 * 1024 * 1024);
-        value = ui::StringUtil::Printf(_T("%.01lf GB"), total_gb);
-    }
-    else if (nSpace > 1 * 1024 * 1024) {
-        //MB
-        double total_mb = static_cast<double>(nSpace) / (1024 * 1024);
-        value = ui::StringUtil::Printf(_T("%.01lf MB"), total_mb);
-    }
-    else if (nSpace > 1 * 1024) {
-        //KB
-        double total_kb = static_cast<double>(nSpace) / (1024);
-        value = ui::StringUtil::Printf(_T("%.01lf KB"), total_kb);
-    }
-    else if (nSpace == 0) {
-        value = _T("0");
-    }
-    else {
-        //B
-        value = ui::StringUtil::Printf(_T("%d B"), (int32_t)nSpace);
-    }
-    return value;
-}
-
-DString MainForm::FormatUsedPercent(uint64_t nTotalSpace, uint64_t nFreeSpace) const
-{
-    DString value;
-    if ((nFreeSpace <= nTotalSpace) && (nTotalSpace != 0)) {
-        double fPercent = static_cast<double>(nTotalSpace - nFreeSpace) / nTotalSpace;
-        value = ui::StringUtil::Printf(_T("%.01lf%%"), fPercent * 100);
-    }
-    return value;
-}
-
-bool MainForm::OnComuterViewDoubleClick(const ui::EventArgs& msg)
-{
-    if ((m_pComputerListCtrl != nullptr) && (msg.wParam != 0) && (m_pTree != nullptr) && (m_pTreeNode != nullptr)) {
-        size_t nItemIndex = msg.lParam;
-        size_t nIndex = m_pComputerListCtrl->GetDataItemUserData(nItemIndex);
-        if (nIndex < m_diskInfoList.size()) {
-            const ui::DirectoryTree::DiskInfo& diskInfo = m_diskInfoList[nIndex];
-            if (!diskInfo.m_filePath.IsEmpty() && diskInfo.m_filePath.IsExistsDirectory() && m_pTree->IsMyComputerNode(m_pTreeNode)) {
-                //进入所选的目录
-                m_pTree->SelectSubPath(m_pTreeNode, diskInfo.m_filePath, nullptr);
-            }
-        }
-    }
-    return true;
-}
-
-void MainForm::ClearDiskInfoList(const std::vector<ui::DirectoryTree::DiskInfo>& diskInfoList) const
-{
-    for (const ui::DirectoryTree::DiskInfo& diskInfo : diskInfoList) {
-        if (!diskInfo.m_bIconShared) {            
-            ui::GlobalManager::Instance().Icon().RemoveIcon(diskInfo.m_nIconID);
-        }
-    }
-}
-
-size_t MainForm::GetColumnId(ComputerViewColumn nOriginIndex) const
-{
-    size_t nColumnIndex = ui::Box::InvalidIndex;
-    auto iter = m_columnIdMap.find(nOriginIndex);
-    if (iter != m_columnIdMap.end()) {
-        return iter->second;
-    }
-    return nColumnIndex;
-}
-
-void MainForm::OnRemoveIcon(uint32_t nIconId)
-{
-    if (!ui::GlobalManager::Instance().IsInUIThread()) {
-        //如果不是在主线程中执行，则转到主线程去执行
-        ui::GlobalManager::Instance().Thread().PostTask(ui::kThreadUI, this->ToWeakCallback([this, nIconId]() {
-                OnRemoveIcon(nIconId);
-            }));
-        return;
-    }
-
-    //在主线程中执行
-    ui::GlobalManager::Instance().AssertUIThread();
-    auto iter = m_iconToImageMap.find(nIconId);
-    if (iter != m_iconToImageMap.end()) {
-        int32_t nImageId = iter->second;
-        m_iconToImageMap.erase(iter);
-
-        ui::ImageListPtr pImageList;
-        if (m_pComputerListCtrl != nullptr) {
-            pImageList = m_pComputerListCtrl->GetImageList(ui::ListCtrlType::Report);
-        }
-        if (pImageList != nullptr) {
-            pImageList->RemoveImageString(nImageId);
-        }
     }
 }
